moved getCountries method from component to service, so that it will not be called with every component contruction but rather once at service initialisation
